adds level

+++ src/C++/DTOs.hpp
@@ -36,7 +36,7 @@ class GetSettingsDto:public oatpp::DTO {
     DTO_FIELD(Int32, duration_slot2);
     DTO_FIELD(Int32, duration_slot3);
     DTO_FIELD(Int32 , lastWater);
-    DTO_FIELD(Int32 , level);
+    DTO_FIELD(Float64 , level);
 
 };
